Divide ParserTest into multi files
